+++ compiler-core/src/parse.rs
@@ -230,9 +230,9 @@ where
 
         let def = match (self.tok0.take(), self.tok1.as_ref()) {
             // Imports
-            (Some((_, Token::Import, _)), _) => {
+            (Some((start, Token::Import, _)), _) => {
                 let _ = self.next_tok();
-                self.parse_import()
+                self.parse_import(start)
             }
             // Module Constants
             (Some((_, Token::Const, _)), _) => {
@@ -924,32 +924,78 @@ where
 
             Some((start, Token::String { value }, end)) => {
                 let _ = self.next_tok();
-                let location = SrcSpan { start, end };
 
-                // String prefix matching
-                // "Hello, " <> name -> ...
-                if let Some((_, Token::LtGt, _)) = self.tok0 {
-                    let _ = self.next_tok();
-                    let (r_start, right, r_end) = self.expect_assign_name()?;
+                match self.tok0 {
+                    // String matching with assignment, it could either be a
+                    // String prefix matching: "Hello, " as greeting <> name -> ...
+                    // or a full string matching: "Hello, World!" as greeting -> ...
+                    Some((_, Token::As, _)) => {
+                        let _ = self.next_tok();
+                        let (name_start, name, name_end) = self.expect_name()?;
+                        let name_span = SrcSpan {
+                            start: name_start,
+                            end: name_end,
+                        };
 
-                    Pattern::Concatenate {
-                        location: SrcSpan {
-                            start: location.start,
-                            end: r_end,
-                        },
-                        left_location: location,
-                        right_location: SrcSpan {
-                            start: r_start,
-                            end: r_end,
-                        },
-                        left_side_string: value,
-                        right_side_assignment: right,
+                        match self.tok0 {
+                            // String prefix matching with assignment
+                            // "Hello, " as greeting <> name -> ...
+                            Some((_, Token::LtGt, _)) => {
+                                let _ = self.next_tok();
+                                let (r_start, right, r_end) = self.expect_assign_name()?;
+                                Pattern::Concatenate {
+                                    location: SrcSpan { start, end: r_end },
+                                    left_location: SrcSpan {
+                                        start,
+                                        end: name_end,
+                                    },
+                                    right_location: SrcSpan {
+                                        start: r_start,
+                                        end: r_end,
+                                    },
+                                    left_side_string: value,
+                                    left_side_assignment: Some((name, name_span)),
+                                    right_side_assignment: right,
+                                }
+                            }
+                            // Full string matching with assignment
+                            _ => {
+                                return Ok(Some(Pattern::Assign {
+                                    name,
+                                    location: name_span,
+                                    pattern: Box::new(Pattern::String {
+                                        location: SrcSpan { start, end },
+                                        value,
+                                    }),
+                                }));
+                            }
+                        }
                     }
 
-                // Full string matching
-                // "Hello, World!" -> ...
-                } else {
-                    Pattern::String { location, value }
+                    // String prefix matching with no left side assignment
+                    // "Hello, " <> name -> ...
+                    Some((_, Token::LtGt, _)) => {
+                        let _ = self.next_tok();
+                        let (r_start, right, r_end) = self.expect_assign_name()?;
+                        Pattern::Concatenate {
+                            location: SrcSpan { start, end: r_end },
+                            left_location: SrcSpan { start, end },
+                            right_location: SrcSpan {
+                                start: r_start,
+                                end: r_end,
+                            },
+                            left_side_string: value,
+                            left_side_assignment: None,
+                            right_side_assignment: right,
+                        }
+                    }
+
+                    // Full string matching
+                    // "Hello, World!" -> ...
+                    _ => Pattern::String {
+                        location: SrcSpan { start, end },
+                        value,
+                    },
                 }
             }
             Some((start, Token::Int { value }, end)) => {
@@ -1888,7 +1934,7 @@ where
     //   import a/b
     //   import a/b.{c}
     //   import a/b.{c as d} as e
-    fn parse_import(&mut self) -> Result<Option<UntypedDefinition>, ParseError> {
+    fn parse_import(&mut self, import_start: u32) -> Result<Option<UntypedDefinition>, ParseError> {
         let mut start = 0;
         let mut end;
         let mut module = String::new();
@@ -1941,7 +1987,10 @@ where
 
         Ok(Some(Definition::Import(Import {
             documentation,
-            location: SrcSpan { start, end },
+            location: SrcSpan {
+                start: import_start,
+                end,
+            },
             unqualified,
             module: module.into(),
             as_name,