+++ test/javascript_prelude/main.mjs
@@ -47,6 +47,15 @@ function assertNotEqual(a, b) {
   }
 }
 
+function assertThrows(msg, callable) {
+  try {
+    callable();
+    fail(msg);
+  } catch (error) {
+    pass();
+  }
+}
+
 class ExampleRecordImpl extends CustomType {
   constructor(first, detail, boop) {
     super();
@@ -602,6 +611,12 @@ assertEqual(
   new ExampleRecordImpl(1, 2, 3).withFields({ boop: 4, detail: 5, 0: 6 }),
   new ExampleRecordImpl(6, 5, 4)
 );
+
+// Test BitString can only be constructed from Uint8Array, not ArrayBuffer
+const bs1 = new BitString(new Uint8Array(new ArrayBuffer(8)));
+assertThrows("Should only construct bitstring from Uint8Array", () => {
+  const bs = new BitString(new ArrayBuffer(8));
+});
 
 //
 // Summary